Reorganize files, components and relative paths
