+++ my-app/.eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/mariacristinamotoc/Desktop/ITAcademy/React-III/my-app/src/index.js":"1","/Users/mariacristinamotoc/Desktop/ITAcademy/React-III/my-app/src/App.js":"2"},{"size":152,"mtime":1611740818000,"results":"3","hashOfConfig":"4"},{"size":2261,"mtime":1612101029000,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"8"},"utugc9",{"filePath":"9","messages":"10","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/Users/mariacristinamotoc/Desktop/ITAcademy/React-III/my-app/src/index.js",[],["11","12"],"/Users/mariacristinamotoc/Desktop/ITAcademy/React-III/my-app/src/App.js",["13"],{"ruleId":"14","replacedBy":"15"},{"ruleId":"16","replacedBy":"17"},{"ruleId":"18","severity":1,"message":"19","line":5,"column":1,"nodeType":"20","endLine":90,"endColumn":3},"no-native-reassign",["21"],"no-negated-in-lhs",["22"],"import/no-anonymous-default-export","Assign arrow function to a variable before exporting as module default","ExportDefaultDeclaration","no-global-assign","no-unsafe-negation"]
+[{"/Users/mariacristinamotoc/Desktop/ITAcademy/React-III/my-app/src/index.js":"1","/Users/mariacristinamotoc/Desktop/ITAcademy/React-III/my-app/src/App.js":"2","/Users/mariacristinamotoc/Desktop/ITAcademy/React-III/my-app/src/components/Jokes.js":"3","/Users/mariacristinamotoc/Desktop/ITAcademy/React-III/my-app/src/components/Weather.js":"4"},{"size":152,"mtime":1611740818000,"results":"5","hashOfConfig":"6"},{"size":698,"mtime":1612103343000,"results":"7","hashOfConfig":"6"},{"size":1139,"mtime":1612103521000,"results":"8","hashOfConfig":"6"},{"size":1013,"mtime":1612103307000,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},"utugc9",{"filePath":"13","messages":"14","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"15","usedDeprecatedRules":"12"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"18","messages":"19","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"20","usedDeprecatedRules":"12"},"/Users/mariacristinamotoc/Desktop/ITAcademy/React-III/my-app/src/index.js",[],["21","22"],"/Users/mariacristinamotoc/Desktop/ITAcademy/React-III/my-app/src/App.js",["23"],"import React from \"react\";\nimport { Container } from \"bootstrap-4-react\";\nimport Jokes from \"./components/Jokes\";\nimport Weather from \"./components/Weather\";\n\nexport default () => {\n\n  // Set Background Image\n  const backgroundImg = {\n    background:\n      \"url('https://i.postimg.cc/Twhcyp8B/pexels-helena-lopes-708440.jpg') no-repeat\",\n    backgroundSize: \"cover\",\n    height: \"100vh\",\n    minHeight: \"100vh\",\n  };\n\n  return (\n    // Added the background image, title, the Jokes and Weather components\n\n    <div style={backgroundImg}>\n\n      <Container>\n        <h1 className=\"text-center text-info\">Web D'Acudits</h1>\n        <Weather />\n        <Jokes />\n      </Container>\n\n    </div>\n  );\n};\n","/Users/mariacristinamotoc/Desktop/ITAcademy/React-III/my-app/src/components/Jokes.js",["24","25"],"/Users/mariacristinamotoc/Desktop/ITAcademy/React-III/my-app/src/components/Weather.js",["26"],"import React, { useState, useEffect } from \"react\";\nimport { Container } from \"bootstrap-4-react\";\nimport axios from \"axios\";\n\nexport default () => {\n  // The Weather information API will be displayed in the beginning. Use axios and useEffect\n\n  const [getWeather, setWeather] = useState({});\n\n  useEffect(() => {\n    axios\n      .get(\n        \"https://api.openweathermap.org/data/2.5/weather?q=Barcelona,es&units=metric&lang=ca&APPID=e50d99da7c6cb79a777875ea8f0a5b6e\"\n      )\n\n      .then((response) => {\n        setWeather(response.data);\n      });\n  }, []);\n\n  return (\n    // Added the weather information: city name, temperature, and weather description. Added Math.round method for temperature. \n\n    <Container>\n      <div className=\" text-left\">\n        {getWeather.main && (\n          <div>\n            <h5>{getWeather.name}</h5>\n            {Math.round(getWeather.main.temp)} º C\n            <p>{getWeather.weather[0].description} </p>\n          </div>\n        )}\n      </div>\n    </Container>\n  );\n};\n",{"ruleId":"27","replacedBy":"28"},{"ruleId":"29","replacedBy":"30"},{"ruleId":"31","severity":1,"message":"32","line":6,"column":1,"nodeType":"33","endLine":30,"endColumn":3},{"ruleId":"34","severity":1,"message":"35","line":1,"column":27,"nodeType":"36","messageId":"37","endLine":1,"endColumn":36},{"ruleId":"31","severity":1,"message":"32","line":5,"column":1,"nodeType":"33","endLine":46,"endColumn":3},{"ruleId":"31","severity":1,"message":"32","line":5,"column":1,"nodeType":"33","endLine":36,"endColumn":3},"no-native-reassign",["38"],"no-negated-in-lhs",["39"],"import/no-anonymous-default-export","Assign arrow function to a variable before exporting as module default","ExportDefaultDeclaration","no-unused-vars","'useEffect' is defined but never used.","Identifier","unusedVar","no-global-assign","no-unsafe-negation"]
